Introduce debug logging of "optimal" delay.
The algorithm to derive the "optimal" delay is very rough, basically
just 95 percentile with largest values cut off. The current results are
not practical because using them would result in a lot more sync events
when using debounce-delay 300 (milliseconds). For now it's just an idea
of something could be become useful if I put thought into it.